Make DirSection shared
