+++ src/lib.rs
@@ -17,7 +17,11 @@ cfg_if::cfg_if! {
 pub mod minidump_cpu;
 pub mod minidump_format;
 
-/// Non-windows platforms need additional code since they are essentially
-/// replicating functionality we get for free on Windows
-#[cfg(not(target_os = "windows"))]
-pub(crate) mod mem_writer;
+// Non-windows platforms need additional code since they are essentially
+// replicating functionality we get for free on Windows
+cfg_if::cfg_if! {
+    if #[cfg(not(target_os = "windows"))] {
+        pub(crate) mod mem_writer;
+        pub(crate) mod dir_section;
+    }
+}
