+++ src/mac/errors.rs
@@ -2,10 +2,10 @@ use thiserror::Error;
 
 #[derive(Debug, Error)]
 pub enum WriterError {
-    #[error("unable to find a UUID for a module")]
-    UnknownUuid,
-    #[error("unable to find the main executable image for the process")]
-    NoExecutableImage,
     #[error(transparent)]
     TaskDumpError(#[from] crate::mac::task_dumper::TaskDumpError),
+    #[error("Failed to write to memory")]
+    MemoryWriterError(#[from] crate::mem_writer::MemoryWriterError),
+    #[error("Failed to write to file")]
+    FileWriterError(#[from] crate::dir_section::FileWriterError),
 }
